fixing castings... (#12)

+++ components/sensors/src/bme280.cpp
@@ -90,26 +90,26 @@ namespace sensors {
         }
 
         std::lock_guard lock(dataMutex);
-        cali.dig_T1 = (uint16_t)(tp_buffer[1] << 8 | tp_buffer[0]);
-        cali.dig_T2 = (int16_t)(tp_buffer[3] << 8 | tp_buffer[2]);
-        cali.dig_T3 = (int16_t)(tp_buffer[5] << 8 | tp_buffer[4]);
+        cali.dig_T1 = static_cast<uint16_t>(tp_buffer[1]) << 8 | tp_buffer[0];
+        cali.dig_T2 = static_cast<int16_t>(tp_buffer[3]) << 8 | tp_buffer[2];
+        cali.dig_T3 = static_cast<int16_t>(tp_buffer[5]) << 8 | tp_buffer[4];
 
-        cali.dig_P1 = (uint16_t)(tp_buffer[7] << 8 | tp_buffer[6]);
-        cali.dig_P2 = (int16_t)(tp_buffer[9] << 8 | tp_buffer[8]);
-        cali.dig_P3 = (int16_t)(tp_buffer[11] << 8 | tp_buffer[10]);
-        cali.dig_P4 = (int16_t)(tp_buffer[13] << 8 | tp_buffer[12]);
-        cali.dig_P5 = (int16_t)(tp_buffer[15] << 8 | tp_buffer[14]);
-        cali.dig_P6 = (int16_t)(tp_buffer[17] << 8 | tp_buffer[16]);
-        cali.dig_P7 = (int16_t)(tp_buffer[19] << 8 | tp_buffer[18]);
-        cali.dig_P8 = (int16_t)(tp_buffer[21] << 8 | tp_buffer[20]);
-        cali.dig_P9 = (int16_t)(tp_buffer[23] << 8 | tp_buffer[22]);
+        cali.dig_P1 = static_cast<uint16_t>(tp_buffer[7]) << 8 | tp_buffer[6];
+        cali.dig_P2 = static_cast<int16_t>(tp_buffer[9]) << 8 | tp_buffer[8];
+        cali.dig_P3 = static_cast<int16_t>(tp_buffer[11]) << 8 | tp_buffer[10];
+        cali.dig_P4 = static_cast<int16_t>(tp_buffer[13]) << 8 | tp_buffer[12];
+        cali.dig_P5 = static_cast<int16_t>(tp_buffer[15]) << 8 | tp_buffer[14];
+        cali.dig_P6 = static_cast<int16_t>(tp_buffer[17]) << 8 | tp_buffer[16];
+        cali.dig_P7 = static_cast<int16_t>(tp_buffer[19]) << 8 | tp_buffer[18];
+        cali.dig_P8 = static_cast<int16_t>(tp_buffer[21]) << 8 | tp_buffer[20];
+        cali.dig_P9 = static_cast<int16_t>(tp_buffer[23]) << 8 | tp_buffer[22];
 
         cali.dig_H1 = h1_buffer[0];
-        cali.dig_H2 = (int16_t)(h2h6_buffer[1] << 8 | h2h6_buffer[0]);
+        cali.dig_H2 = static_cast<int16_t>(h2h6_buffer[1]) << 8 | h2h6_buffer[0];
         cali.dig_H3 = h2h6_buffer[2];
-        cali.dig_H4 = (int16_t)(h2h6_buffer[3] << 4 | (h2h6_buffer[4] & 0x0F));
-        cali.dig_H5 = (int16_t)(h2h6_buffer[5] << 4 | h2h6_buffer[4] >> 4);
-        cali.dig_H6 = (int8_t)h2h6_buffer[6];
+        cali.dig_H4 = static_cast<int16_t>(h2h6_buffer[3]) << 4 | (h2h6_buffer[4] & 0x0F);
+        cali.dig_H5 = static_cast<int16_t>(h2h6_buffer[5]) << 4 | (h2h6_buffer[4] >> 4);
+        cali.dig_H6 = h2h6_buffer[6];
     }
 
     void BME280::read_all() {
@@ -122,26 +122,28 @@ namespace sensors {
         // Register 1 = Most Significant byte (use all 8 bits)
         // Register 2 = Least Significant byte (use all 8 bits)
         // Register 3 = Extra Least Significant bits (use only left 4 bits)
-        uint32_t raw_press = (buffer[0] << 12) | buffer[1] << 4 | buffer[2] >> 4;
+        uint32_t raw_press = (static_cast<uint32_t>(buffer[0]) << 12) |
+                             static_cast<uint32_t>(buffer[1]) << 4 | buffer[2] >> 4;
 
         // The next three registers are for Temperature (20 bits total).
         // Register 1 = Most Significant byte (use all 8 bits)
         // Register 2 = Least Significant byte (use all 8 bits)
         // Register 3 = Extra Least Significant bits (use only left 4 bits)
-        uint32_t raw_temp = (buffer[3] << 12) | buffer[4] << 4 | buffer[5] >> 4;
+        uint32_t raw_temp = (static_cast<uint32_t>(buffer[3]) << 12) |
+                            static_cast<uint32_t>(buffer[4]) << 4 | buffer[5] >> 4;
 
         // The last two registers are for Humidity (16 bits total).
         // Register 1 = Most Significant byte (use all 8 bits)
         // Register 2 = Least Significant byte (use all 8 bits)
-        uint32_t raw_hum = (buffer[6] << 8) | buffer[7];
+        uint32_t raw_hum = (static_cast<uint32_t>(buffer[6]) << 8) | buffer[7];
 
         // Compensation formulas from BME280 datasheet
         int32_t t_fine = solve_tfine(raw_temp);
 
         std::lock_guard lock(dataMutex);
-        temperature = (float)((t_fine * 5 + 128) >> 8) / 100.0;
+        temperature = static_cast<float>((t_fine * 5 + 128) >> 8) / 100.0f;
         pressure = compensate_pressure(raw_press, t_fine);
-        humidity = (double)compensate_humidity(raw_hum, t_fine) / 1024.0;
+        humidity = static_cast<double>(compensate_humidity(raw_hum, t_fine)) / 1024.0;
 
         ESP_LOGV(name.c_str(), "Temperature : %f", temperature);
         ESP_LOGV(name.c_str(), "Pressure: %f", pressure);
@@ -149,57 +151,60 @@ namespace sensors {
     }
 
     int32_t BME280::solve_tfine(uint32_t raw_temp) {
-        int32_t var1 =
-            ((((raw_temp >> 3) - ((int32_t)cali.dig_T1 << 1))) * ((int32_t)cali.dig_T2)) >> 11;
-        int32_t var2 = (((((raw_temp >> 4) - ((int32_t)cali.dig_T1)) *
-                          ((raw_temp >> 4) - ((int32_t)cali.dig_T1))) >>
+        int32_t var1 = ((((raw_temp >> 3) - (static_cast<int32_t>(cali.dig_T1) << 1))) *
+                        static_cast<int32_t>(cali.dig_T2)) >>
+                       11;
+
+        int32_t var2 = (((((raw_temp >> 4) - static_cast<int32_t>(cali.dig_T1)) *
+                          ((raw_temp >> 4) - static_cast<int32_t>(cali.dig_T1))) >>
                          12) *
-                        ((int32_t)cali.dig_T3)) >>
+                        static_cast<int32_t>(cali.dig_T3)) >>
                        14;
         int32_t t_fine = var1 + var2;
         return t_fine;
     }
 
     double BME280::compensate_pressure(uint32_t raw_pressure, int32_t t_fine) {
-        double var1 = ((double)t_fine / 2.0) - 64000.0;
-        double var2 = var1 * var1 * ((double)cali.dig_P6) / 32768.0;
-        var2 = var2 + var1 * ((double)cali.dig_P5) * 2.0;
-        var2 = (var2 / 4.0) + (((double)cali.dig_P4) * 65536.0);
-        var1 = (((double)cali.dig_P3) * var1 * var1 / 524288.0 + ((double)cali.dig_P2) * var1) /
+        double var1 = (static_cast<double>(t_fine) / 2.0) - 64000.0;
+        double var2 = var1 * var1 * static_cast<double>(cali.dig_P6) / 32768.0;
+        var2 = var2 + var1 * static_cast<double>(cali.dig_P5) * 2.0;
+        var2 = (var2 / 4.0) + (static_cast<double>(cali.dig_P4) * 65536.0);
+        var1 = (static_cast<double>(cali.dig_P3) * var1 * var1 / 524288.0 +
+                static_cast<double>(cali.dig_P2) * var1) /
                524288.0;
-        var1 = (1.0 + var1 / 32768.0) * ((double)cali.dig_P1);
+        var1 = (1.0 + var1 / 32768.0) * static_cast<double>(cali.dig_P1);
         if (var1 == 0) {
             return 0;  // Avoid dividing by 0
         }
-        double p = 1048576.0 - (double)raw_pressure;
+        double p = 1048576.0 - static_cast<double>(raw_pressure);
         p = (p - (var2 / 4096.0)) * 6250.0 / var1;
-        var1 = ((double)cali.dig_P9) * p * p / 2147483648.0;
-        var2 = p * ((double)cali.dig_P8) / 32768.0;
-        p = p + (var1 + var2 + ((double)cali.dig_P7)) / 16.0;
+        var1 = static_cast<double>(cali.dig_P9) * p * p / 2147483648.0;
+        var2 = p * static_cast<double>(cali.dig_P8) / 32768.0;
+        p = p + (var1 + var2 + static_cast<double>(cali.dig_P7)) / 16.0;
         return p;
     }
 
     uint32_t BME280::compensate_humidity(uint32_t raw_humidity, int32_t t_fine) {
-        int32_t v_x1_u32r = t_fine - (int32_t)76800;
-        v_x1_u32r = (((((raw_humidity << 14) - (((int32_t)cali.dig_H4) << 20) -
-                        (((int32_t)cali.dig_H5) * v_x1_u32r)) +
-                       ((int32_t)16384)) >>
+        int32_t v_x1_u32r = t_fine - 76800;
+        v_x1_u32r = (((((raw_humidity << 14) - (static_cast<int32_t>(cali.dig_H4) << 20) -
+                        (static_cast<int32_t>(cali.dig_H5) * v_x1_u32r)) +
+                       16384) >>
                       15) *
-                     (((((((v_x1_u32r * ((int32_t)cali.dig_H6)) >> 10) *
-                          (((v_x1_u32r * ((int32_t)cali.dig_H3)) >> 11) + ((int32_t)32768))) >>
+                     (((((((v_x1_u32r * static_cast<int32_t>(cali.dig_H6)) >> 10) *
+                          (((v_x1_u32r * static_cast<int32_t>(cali.dig_H3)) >> 11) + 32768)) >>
                          10) +
-                        ((int32_t)2097152)) *
-                           ((int32_t)cali.dig_H2) +
+                        2097152) *
+                           static_cast<int32_t>(cali.dig_H2) +
                        8192) >>
                       14));
 
-        v_x1_u32r =
-            (v_x1_u32r -
-             (((((v_x1_u32r >> 15) * (v_x1_u32r >> 15)) >> 7) * ((int32_t)cali.dig_H1)) >> 4));
+        v_x1_u32r = (v_x1_u32r - (((((v_x1_u32r >> 15) * (v_x1_u32r >> 15)) >> 7) *
+                                   static_cast<int32_t>(cali.dig_H1)) >>
+                                  4));
         v_x1_u32r = (v_x1_u32r < 0 ? 0 : v_x1_u32r);
         v_x1_u32r = (v_x1_u32r > 419430400 ? 419430400 : v_x1_u32r);
 
-        return (uint32_t)(v_x1_u32r >> 12);
+        return (v_x1_u32r >> 12);
     }
 
     uint32_t BME280::getHumidity() const {
